refactor: move mem/ to src/ top-level module
