+++ src/hooks/mod.rs
@@ -3,13 +3,10 @@ pub mod config;
 pub mod guard;
 pub mod install;
 mod install_wizard;
-pub mod mem;
 pub mod observe;
-pub mod orchestrate;
 pub mod polish;
 pub mod reflect;
 pub mod resume;
 pub mod serve;
 pub mod snapshot;
-pub mod team;
 pub mod telemetry;
