+++ src/hooks/resume.rs
@@ -3,7 +3,7 @@ use std::io::ErrorKind;
 use std::path::Path;
 
 use super::common::*;
-use super::mem::store;
+use crate::mem::store;
 use super::telemetry::Telemetry;
 
 /// Atomically acquire a session lock file.
@@ -217,8 +217,8 @@ pub fn run(_input: &HookInput) -> i32 {
     }
 
     // Seed the default org/team whenever orgs dir is empty (idempotent)
-    let default_org = crate::hooks::team::store::default_org();
-    if crate::hooks::team::store::install_default_team_if_needed(&default_org) {
+    let default_org = crate::team::store::default_org();
+    if crate::team::store::install_default_team_if_needed(&default_org) {
         hint(
             "resume",
             &format!(
