+++ src/mem/cli.rs
