+++ src/mem/store/index.rs
