+++ src/mem/store/node.rs
