+++ src/mem/store/search.rs
