+++ src/mem/store/tests.rs
