+++ src/mem/store/types.rs
