+++ src/mem/store/util.rs
