+++ src/team/cli.rs
